Refector : enum entity에 병합, column에 type 글씨 추가
