--- src/jobs/enum/job- payment.enum.ts
@@ -1,8 +1,0 @@
-export enum PayMentsMethod {
-	/* 시급지불 */
-	PERHOUR = "PERHOUR",
-	/* 일당지불 */
-	PERDAY = "PERDAY",
-	/* 월급 지불 */
-	PERMONTH = "PERMONTH",
-}
