--- src/jobs/enum/job-status.enum.ts
@@ -1,6 +1,0 @@
-export enum JobStatus {
-	// 모집중
-	ACTIVATE = "ACTIVATE",
-	/* 모집 완료 */
-	INAVCTIVE = "INAVCTIVE",
-}
